+++ drivers/display/CMakeLists.txt
@@ -6,7 +6,6 @@ zephyr_library_sources_ifdef(CONFIG_DISPLAY_NRF_LED_MATRIX display_nrf_led_matri
 zephyr_library_sources_ifdef(CONFIG_DUMMY_DISPLAY	display_dummy.c)
 zephyr_library_sources_ifdef(CONFIG_FRAMEBUF_DISPLAY	display_framebuf.c)
 zephyr_library_sources_ifdef(CONFIG_GD7965		gd7965.c)
-zephyr_library_sources_ifdef(CONFIG_GROVE_LCD_RGB	grove_lcd_rgb.c)
 zephyr_library_sources_ifdef(CONFIG_ILI9XXX		display_ili9xxx.c)
 zephyr_library_sources_ifdef(CONFIG_ILI9340		display_ili9340.c)
 zephyr_library_sources_ifdef(CONFIG_ILI9341		display_ili9341.c)
