+++ drivers/display/Kconfig
@@ -20,7 +20,6 @@ module = DISPLAY
 module-str = display
 source "subsys/logging/Kconfig.template.log_config"
 
-source "drivers/display/Kconfig.grove"
 source "drivers/display/Kconfig.mcux_elcdif"
 source "drivers/display/Kconfig.microbit"
 source "drivers/display/Kconfig.nrf_led_matrix"
