+++ drivers/misc/grove_lcd_rgb/CMakeLists.txt
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+
+zephyr_library_sources(grove_lcd_rgb.c)
