+++ drivers/misc/grove_lcd_rgb/Kconfig
@@ -1,3 +1,4 @@
+# Copyright (c) 2015 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 config GROVE_LCD_RGB
