+++ samples/drivers/misc/grove_display/prj.conf
@@ -3,5 +3,4 @@ CONFIG_PRINTK=y
 
 CONFIG_I2C=y
 
-CONFIG_DISPLAY=y
 CONFIG_GROVE_LCD_RGB=y
