+++ samples/drivers/misc/grove_display/sample.yaml
@@ -1,7 +1,7 @@
 sample:
   name: Grove LCD Sample
 tests:
-  sample.display.grove_display:
+  sample.drivers.misc.grove_display:
     tags: drivers
     harness: grove
     depends_on: i2c
