+++ samples/sensor/grove_temperature/src/main.c
@@ -10,7 +10,7 @@
 #include <drivers/sensor.h>
 
 #ifdef CONFIG_GROVE_LCD_RGB
-#include <display/grove_lcd.h>
+#include <drivers/misc/grove_lcd/grove_lcd.h>
 #include <stdio.h>
 #include <string.h>
 #endif
