+++ samples/sensor/th02/prj.conf
@@ -5,5 +5,4 @@ CONFIG_TH02=y
 CONFIG_LOG=n
 CONFIG_SENSOR_LOG_LEVEL_DBG=y
 CONFIG_GROVE_LCD_RGB=y
-CONFIG_DISPLAY=y
 CONFIG_CBPRINTF_FP_SUPPORT=y
